refactor sp tests package
